More Wconversion errors resolved.

+++ src_gap/Combinatorics.h
@@ -14,15 +14,15 @@ void SortParallel_PairList(std::vector<T1> & list1, std::vector<T2> & list2)
     T1 val1;
     T2 val2;
   };
-  int len=list1.size();
+  size_t len=list1.size();
   std::vector<PairData> ListPair(len);
-  for (int i=0; i<len; i++)
+  for (size_t i=0; i<len; i++)
     ListPair[i] = {list1[i], list2[i]};
   sort(ListPair.begin(), ListPair.end(),
        [](PairData const & a, PairData const& b) -> bool {
          return a.val1 < b.val1;
        });
-  for (int i=0; i<len; i++) {
+  for (size_t i=0; i<len; i++) {
     list1[i] = ListPair[i].val1;
     list2[i] = ListPair[i].val2;
   }
